+++ comet_pqc/plugins/summary.py
@@ -61,9 +61,9 @@ class SummaryPlugin:
         self.window.removePage(self.summaryWidget)
         self.summaryWidget.deleteLater()
 
-    def on_summary(self, data: dict) -> None:
+    def on_measurement_finished(self, data: dict) -> None:
         """Push result to summary and write to summary file (experimantal)."""
-        item = self.summaryWidget.appendResult(data)
+        item = self.summaryWidget.appendResult(data)  # TODO
         output_path = self.window.dashboard.outputDir()
         if output_path and os.path.exists(output_path):
             filename = os.path.join(output_path, SUMMARY_FILENAME)
